% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{boot.wild}
\alias{boot.wild}
\title{Title boot.wild}
\usage{
boot.wild(tau, n, B.num, weight = NULL, result, design_x)
}
\arguments{
\item{tau}{quantile level}

\item{n}{sample size}

\item{B.num}{wildbootstrap number}

\item{weight}{weight}

\item{result}{quantile regression result}

\item{design_x}{design matrix}
}
\value{
a matrix
}
\description{
Title boot.wild
}
\examples{
cov.boot = boot.wild(tau,n,B.num,weight = NULL,result,design_x)
}
