% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{cal_B}
\alias{cal_B}
\title{Title cal_B}
\usage{
cal_B(q, p)
}
\arguments{
\item{q}{order of polynomial function}

\item{p}{number of spline basis functions}

\item{kjn}{knots number of spline}
}
\value{
list
}
\description{
Title cal_B
}
\examples{
cj1 = cal_B(q = 2,p = 8,kjn = 5)[[1]]
}
