% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{cal_C}
\alias{cal_C}
\title{Title cal_C}
\usage{
cal_C(p, D, J, ind)
}
\arguments{
\item{p}{number of spline basis functions}

\item{D}{number of linear parameters}

\item{J}{number of additive functions}

\item{ind}{which additive component you want to test}
}
\value{
matrix
}
\description{
Title cal_C
}
\examples{
C = cal_C(p = 8,D = 3,J = 3,ind = 1)
}
