% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{cal_H}
\alias{cal_H}
\title{Title cal_H}
\usage{
cal_H(p, q, J0)
}
\arguments{
\item{p}{number of spline basis functions}

\item{q}{order of polynomial function}

\item{J0}{number of additive components that you want to test}
}
\value{
list
}
\description{
Title cal_H
}
\examples{
H = cal_H(p = 8,q = 2,J0 = 1)
}
