% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{cal_pvalue}
\alias{cal_pvalue}
\title{Title cal_pvalue}
\usage{
cal_pvalue(r, J0)
}
\arguments{
\item{r}{a list}

\item{J0}{number of additive functions you want to test}
}
\value{
a number
}
\description{
Title cal_pvalue
}
\examples{
pvalue = cal_pvalue(r,J0)
}
