% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{calcu_Bz}
\alias{calcu_Bz}
\title{Title calcu_Bz}
\usage{
calcu_Bz(q, p, degree, z.sp)
}
\arguments{
\item{q}{polynomial order}

\item{p}{number of spline basis functions}

\item{degree}{order of spline}

\item{z.sp}{list}
}
\value{
a matrix
}
\description{
Title calcu_Bz
}
\examples{
B_z = calcu_Bz(q,p,degree,z.sp)
}
