% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{coef.order}
\alias{coef.order}
\title{Title coef.order}
\usage{
\method{coef}{order}(uu1, p, order, degree)
}
\arguments{
\item{uu1}{knots of spline}

\item{p}{number of spline basis functions}

\item{order}{derivative order now calcu :1-q}

\item{degree}{spline order}
}
\value{
vector
}
\description{
Title coef.order
}
\examples{
coef.order1 = coef.order(uu1,p,q,order=1,degree)
}
