% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{kn.sic}
\alias{kn.sic}
\title{Title kn.sic}
\usage{
kn.sic(y, x, z, n, tau, degree, ind)
}
\arguments{
\item{y}{response}

\item{x}{linear covariate}

\item{z}{additive covariate}

\item{n}{sample size}

\item{tau}{quantile level}

\item{degree}{order of spline}

\item{ind}{index indicate which additve function is testing now}
}
\value{
a number
}
\description{
Title kn.sic
}
\examples{
kjn = kn.sic(y,x,z,n,tau,degree,ind = 1)
}
