% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{knots}
\alias{knots}
\title{Title knots}
\usage{
knots(z, u)
}
\arguments{
\item{z}{additive covariate}

\item{u}{sequence u}
}
\value{
vector
}
\description{
Title knots
}
\examples{
knots_z = knots(z,u)
}
