% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{myfit}
\alias{myfit}
\title{Title myfit}
\usage{
myfit(y, design_x, tau)
}
\arguments{
\item{y}{response}

\item{design_x}{covariate x and b-spline basis}

\item{tau}{quantile level}
}
\value{

}
\description{
Title myfit
}
\examples{
result = myfit(y,design_x,tau = 0.5)
}
