% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{Plamct.test}
\alias{Plamct.test}
\title{Title Plamct.test}
\usage{
Plamct.test(y, x, z, n, tau, degree, J, D, q, ind, B.num)
}
\arguments{
\item{y}{response y}

\item{x}{linear covariate}

\item{z}{additve covariate}

\item{n}{sample size}

\item{tau}{quantile level}

\item{degree}{order of spline}

\item{J}{number of additive functions}

\item{D}{number of linear part}

\item{q}{polynomial order}

\item{ind}{index}

\item{B.num}{wildbootstrap number}
}
\value{
a number
}
\description{
Title Plamct.test
}
\examples{
a= Plamct.test(y,x,z,n,tau,degree,J,D,q,ind,B.num)
}
