% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{poly.test}
\alias{poly.test}
\title{Title poly.test}
\usage{
poly.test(
  result,
  design_x,
  tau,
  n,
  J,
  D,
  p,
  q,
  degree,
  B.num,
  weight = NULL,
  z.sp,
  ind
)
}
\arguments{
\item{result}{quantile regression result}

\item{design_x}{x and z_spline}

\item{tau}{quantile level}

\item{n}{sample size}

\item{J}{number of additive functions}

\item{D}{number of linear part}

\item{p}{number of spline basis functions}

\item{q}{polynomial order}

\item{degree}{order of spline}

\item{B.num}{wildbootstrap number}

\item{weight}{wildbootstrap weight}

\item{z.sp}{list}

\item{ind}{index}
}
\value{
a list
}
\description{
Title poly.test
}
\examples{
r = poly.test(result,design_x,tau,n,J,D,p,q,degree,B.num,weight=NULL,z.sp,ind = 1)
}
