% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{score}
\alias{score}
\title{Title score}
\usage{
score(TAU, e_hat)
}
\arguments{
\item{TAU}{quantile level}

\item{e_hat}{estimated error}
}
\value{
a number
}
\description{
Title score
}
\examples{
s = score(tau,result$residuals)
}
