% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{Tn}
\alias{Tn}
\title{Title Tn}
\usage{
Tn(H, A, Bj, gamma, coef_z)
}
\arguments{
\item{H}{matrix H}

\item{A}{matrix A}

\item{Bj}{matrix Bj}

\item{gamma}{matrix gamma}

\item{coef_z}{coefficient of spline}
}
\value{
a number
}
\description{
Title Tn
}
\examples{
stat.norm = Tn(H1,A1,B_z1,gamma1,coef1)
}
