% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{uu1}
\alias{uu1}
\title{Title knots:calcu uu1}
\usage{
uu1(p, degree, boundary.knots, Knots_z1)
}
\arguments{
\item{p}{number of spline basis functions}

\item{degree}{spline order}

\item{boundary.knots}{bourdary.knots of spline}

\item{Knots_z1}{knots_z1}
}
\value{
vector
}
\description{
Title knots:calcu uu1
}
\examples{
uu1 = knots(p,degree,bound,z.knots.z1)
}
