% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{z.basis}
\alias{z.basis}
\title{Title z.basis}
\usage{
z.basis(z, u, ind, p)
}
\arguments{
\item{z}{additive covariate}

\item{u}{vector u}

\item{ind}{index}

\item{p}{number of spline basis functions}
}
\value{
matrix
}
\description{
Title z.basis
}
\examples{
basis_z = z.basis(z,u,ind,p)
}
