% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plamct.R
\name{z.spline}
\alias{z.spline}
\title{Title z.spline}
\usage{
z.spline(y, x, z, n, p, tau, kjn, degree, ind)
}
\arguments{
\item{y}{response}

\item{x}{linear covariate}

\item{z}{additive covariate}

\item{n}{sample size}

\item{p}{number of spline basis functions}

\item{tau}{quantile level}

\item{kjn}{number of spline knots}

\item{degree}{order of spline}

\item{ind}{index}
}
\value{
a list
}
\description{
Title z.spline
}
\examples{
z.sp = z.spline(y,x,z,n,p,tau,kjn,degree,ind = 1)
}
